baitap-submit/dan-dinh/08-crypto-chatbot/requirements.txt
groq==0.15.0
python-dotenv==0.21.0
gradio==5.12.0
pydantic==2.10.5
httpx==0.27.2
yfinance==0.2.55